cloud-functions/model-indexes/requirements.txt
requests==2.32.3
Flask==2.3.3
python-dateutil==2.8.2
numpy==1.24.3
pandas==2.2.0
google-cloud-storage==2.17.0